chore: switch back to archlinux
nixos was nice, solid and clean system, almost no breakage in a few
months.
But developping in C is a hassle with it, linking is quite bad, need to
write shell.nix files everywhere. Not linuxy enough so it feels a bit
like a lost learning time.
This commit fixes the configs and scripts and updates
submodules.
A switch to fast-syntax-highlighting is needed because of conflicts with
zsh old plugin.